wip:Todo機能完了/categoryもほぼ完了

+++ app/lib/persistence/db/ToDoCategory.scala
@@ -38,12 +38,12 @@ case class TodoCategoryTable[P <: JdbcProfile]()(implicit val driver: P)
     /* @1 */ def id         = column[Id]            ("id",          O.UInt64, O.PrimaryKey, O.AutoInc)
     /* @3 */ def name       = column[String]        ("name",        O.Utf8Char255)
     /* @3 */ def slug       = column[String]        ("slug",        O.Utf8Char64)
-    /* @4 */ def color      = column[Int]           ("color",       O.UInt8)
+    /* @4 */ def color      = column[Short]         ("color",       O.UInt8)
     /* @5 */ def updatedAt  = column[LocalDateTime] ("updated_at",  O.TsCurrent)
     /* @6 */ def createdAt  = column[LocalDateTime] ("created_at",  O.Ts)
 
     type TableElementTuple = (
-      Option[Id], String, String, Int, LocalDateTime, LocalDateTime
+      Option[Id], String, String, Short, LocalDateTime, LocalDateTime
       )
 
     // DB <=> Scala の相互のmapping定義
